requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
python-dateutil==2.9.0.post0
colorama==0.4.6
pytest==9.0.1